+++ feature/news/src/main/java/com/teamwable/news/news/NewsNewsRoute.kt
@@ -33,8 +33,8 @@ import com.teamwable.model.news.NewsInfoModel
 import com.teamwable.news.R
 import com.teamwable.news.news.component.WableNewsItems
 import com.teamwable.news.news.component.WableNewsTopBanner
-import com.teamwable.news.news.contract.NewsInfoIntent
-import com.teamwable.news.news.contract.NewsInfoSideEffect
+import com.teamwable.news.news.model.NewsInfoIntent
+import com.teamwable.news.news.model.NewsInfoSideEffect
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.flowOf
