+++ feature/news/src/main/java/com/teamwable/news/notice/NewsNoticeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.cachedIn
 import com.teamwable.data.repository.NewsRepository
 import com.teamwable.model.news.NewsInfoModel
-import com.teamwable.news.news.contract.NewsInfoSideEffect
+import com.teamwable.news.news.model.NewsInfoSideEffect
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
